LayoutTests/fast/innerHTML/javascript-url-expected.txt

Test that innerHTML/outerHTML does not mangle javascript: urls.
PASS r.innerHTML is "<a href=\"javascript:test(123)\">link</a>"
PASS r.innerHTML is "<a href='javascript:test(\"text<\")'>link</a>"
PASS r.innerHTML is "<a href=\"javascript:test('text>')\">link</a>"
PASS r.innerHTML is "<a href=\"javascript:test('text&amp;',&quot;test2&amp;&quot;)\">link</a>"
PASS r.firstChild.outerHTML is "<a href=\"javascript:window.location='?x&amp;y'\">link</a>"
PASS r.firstChild.outerHTML is "<a href=\"javascript:window.location='?x&amp;amp;y'\">link</a>"
PASS r.firstChild.outerHTML is "<a href=\"javascript:window.location='?x&amp;y'\">link</a>"
PASS successfullyParsed is true

TEST COMPLETE
link
